Move advance search to multi-src
