+++ src/all/imhentai/src/eu/kanade/tachiyomi/extension/all/imhentai/IMHentai.kt
@@ -6,14 +6,9 @@ import androidx.preference.SwitchPreferenceCompat
 import eu.kanade.tachiyomi.multisrc.galleryadults.GalleryAdults
 import eu.kanade.tachiyomi.multisrc.galleryadults.GalleryAdultsUtils.cleanTag
 import eu.kanade.tachiyomi.multisrc.galleryadults.GalleryAdultsUtils.imgAttr
-import eu.kanade.tachiyomi.network.GET
-import eu.kanade.tachiyomi.network.POST
-import eu.kanade.tachiyomi.source.model.Filter
-import eu.kanade.tachiyomi.source.model.FilterList
 import okhttp3.FormBody
 import okhttp3.HttpUrl.Companion.toHttpUrl
 import okhttp3.OkHttpClient
-import okhttp3.Request
 import okhttp3.Response
 import okhttp3.ResponseBody.Companion.toResponseBody
 import org.jsoup.nodes.Document
@@ -29,6 +24,9 @@ class IMHentai(
     lang = lang,
 ) {
     override val supportsLatest = true
+    override val useIntermediateSearch: Boolean = true
+    override val useAdvanceSearch: Boolean = true
+    override val supportSpeechless: Boolean = true
 
     private val SharedPreferences.shortTitle
         get() = getBoolean(PREF_SHORT_TITLE, false)
@@ -93,121 +91,6 @@ class IMHentai(
                     .build()
             },
         ).build()
-
-    private fun toBinary(boolean: Boolean) = if (boolean) "1" else "0"
-
-    override fun searchMangaRequest(page: Int, query: String, filters: FilterList): Request {
-        val sortOrderFilter = filters.filterIsInstance<SortOrderFilter>().firstOrNull()
-        val genresFilter = filters.filterIsInstance<GenresFilter>().firstOrNull()
-        val selectedGenres = genresFilter?.state?.filter { it.state } ?: emptyList()
-        val favoriteFilter = filters.filterIsInstance<FavoriteFilter>().firstOrNull()
-        val speechlessFilter = filters.filterIsInstance<SpeechlessFilter>().firstOrNull()
-        val categoryFilters = filters.filterIsInstance<CategoryFilters>().firstOrNull()
-        val advancedSearchFilters = filters.filterIsInstance<Filter.Text>()
-        return when {
-            favoriteFilter?.state == true -> {
-                val url = "$baseUrl/$favoritePath".toHttpUrl().newBuilder()
-                return POST(
-                    url.build().toString(),
-                    xhrHeaders,
-                    FormBody.Builder()
-                        .add("page", page.toString())
-                        .build(),
-                )
-            }
-            speechlessFilter?.state == true -> {
-                val url = baseUrl.toHttpUrl().newBuilder().apply {
-                    addPathSegment("language")
-                    addPathSegment(LANGUAGE_SPEECHLESS)
-                    if (sortOrderFilter?.state == 0) addPathSegment("popular")
-                    addPageUri(page)
-                }
-                GET(url.build(), headers)
-            }
-            advancedSearchFilters.any { it.state.isNotBlank() } -> {
-                val url = "$baseUrl/advsearch".toHttpUrl().newBuilder().apply {
-                    getSortOrderURIs().forEachIndexed { index, pair ->
-                        addQueryParameter(pair.second, toBinary(sortOrderFilter?.state == index))
-                    }
-                    categoryFilters?.state?.forEach {
-                        addQueryParameter(it.uri, toBinary(it.state))
-                    }
-                    getLanguageURIs().forEach { pair ->
-                        addQueryParameter(
-                            pair.second,
-                            toBinary(
-                                mangaLang == pair.first ||
-                                    mangaLang == LANGUAGE_MULTI,
-                            ),
-                        )
-                    }
-
-                    // Build this query string: +tag:"bat+man"+-tag:"cat"+artist:"Joe"...
-                    // +tag must be encoded into %2Btag while the rest are not needed to encode
-                    val keys = emptyList<String>().toMutableList()
-                    keys.addAll(selectedGenres.map { "%2Btag:\"${it.name}\"" })
-                    advancedSearchFilters.forEach { filter ->
-                        val key = when (filter) {
-                            is TagsFilter -> "tag"
-                            is ParodiesFilter -> "parody"
-                            is ArtistsFilter -> "artist"
-                            is CharactersFilter -> "character"
-                            is GroupsFilter -> "group"
-                            else -> null
-                        }
-                        if (key != null) {
-                            keys.addAll(
-                                filter.state.trim()
-                                    // any space except after a comma (we're going to replace spaces only between words)
-                                    .replace(Regex("""(?<!,)\s+"""), "+")
-                                    .replace(" ", "")
-                                    .split(',')
-                                    .mapNotNull {
-                                        val match = Regex("""^(-?)"?(.+)"?""").find(it)
-                                        match?.groupValues?.let { groups ->
-                                            "${if (groups[1].isNotBlank()) "-" else "%2B"}$key:\"${groups[2]}\""
-                                        }
-                                    },
-                            )
-                        }
-                    }
-                    addEncodedQueryParameter("key", keys.joinToString("+"))
-                    addPageUri(page)
-                }
-                GET(url.build())
-            }
-            selectedGenres.size == 1 && query.isBlank() -> {
-                // Browsing single tag's catalog
-                val url = baseUrl.toHttpUrl().newBuilder().apply {
-                    addPathSegment("tag")
-                    addPathSegment(selectedGenres.single().uri)
-                    if (sortOrderFilter?.state == 0) addPathSegment("popular")
-                    addPageUri(page)
-                }
-                GET(url.build(), headers)
-            }
-            else -> {
-                // Only for query string or multiple tags
-                val url = "$baseUrl/search".toHttpUrl().newBuilder().apply {
-                    getSortOrderURIs().forEachIndexed { index, pair ->
-                        addQueryParameter(pair.second, toBinary(sortOrderFilter?.state == index))
-                    }
-                    categoryFilters?.state?.forEach {
-                        addQueryParameter(it.uri, toBinary(it.state))
-                    }
-                    getLanguageURIs().forEach { pair ->
-                        addQueryParameter(
-                            pair.second,
-                            toBinary(mangaLang == pair.first || mangaLang == LANGUAGE_MULTI),
-                        )
-                    }
-                    addEncodedQueryParameter("key", buildQueryString(selectedGenres.map { it.name }, query))
-                    addPageUri(page)
-                }
-                GET(url.build())
-            }
-        }
-    }
 
     override val favoritePath = "user/fav_pags.php"
 
@@ -306,73 +189,6 @@ class IMHentai(
                 )
             }
     }
-
-    private class SortOrderFilter(sortOrderURIs: List<Pair<String, String>>) :
-        Filter.Select<String>("Sort By", sortOrderURIs.map { it.first }.toTypedArray())
-
-    private open class SearchFlagFilter(name: String, val uri: String, state: Boolean = true) : Filter.CheckBox(name, state)
-    private class CategoryFilters(flags: List<SearchFlagFilter>) : Filter.Group<SearchFlagFilter>("Categories", flags)
-
-    private class SpeechlessFilter : Filter.CheckBox("Show speechless items only", false)
-
-    private class TagsFilter : Filter.Text("Tags")
-    private class ParodiesFilter : Filter.Text("Parodies")
-    private class ArtistsFilter : Filter.Text("Artists")
-    private class CharactersFilter : Filter.Text("Characters")
-    private class GroupsFilter : Filter.Text("Groups")
-
-    override fun getFilterList(): FilterList {
-        getGenres()
-        return FilterList(
-            if (genres.isEmpty()) {
-                Filter.Header("Press 'reset' to attempt to load tags")
-            } else {
-                GenresFilter(genres)
-            },
-            Filter.Separator(),
-
-            SortOrderFilter(getSortOrderURIs()),
-            CategoryFilters(getCategoryURIs()),
-
-            Filter.Separator(),
-            Filter.Header("Advanced filters will ignore query search. Separate terms by comma (,) and precede term with minus (-) to exclude."),
-            TagsFilter(),
-            ParodiesFilter(),
-            ArtistsFilter(),
-            CharactersFilter(),
-            GroupsFilter(),
-
-            Filter.Separator(),
-            SpeechlessFilter(),
-            FavoriteFilter(),
-        )
-    }
-
-    private fun getCategoryURIs() = listOf(
-        SearchFlagFilter("Manga", "m"),
-        SearchFlagFilter("Doujinshi", "d"),
-        SearchFlagFilter("Western", "w"),
-        SearchFlagFilter("Image Set", "i"),
-        SearchFlagFilter("Artist CG", "a"),
-        SearchFlagFilter("Game CG", "g"),
-    )
-
-    private fun getSortOrderURIs() = listOf(
-        Pair("Popular", "pp"),
-        Pair("Latest", "lt"),
-        Pair("Downloads", "dl"),
-        Pair("Top Rated", "tr"),
-    )
-
-    private fun getLanguageURIs() = listOf(
-        Pair(LANGUAGE_ENGLISH, "en"),
-        Pair(LANGUAGE_JAPANESE, "jp"),
-        Pair(LANGUAGE_SPANISH, "es"),
-        Pair(LANGUAGE_FRENCH, "fr"),
-        Pair(LANGUAGE_KOREAN, "kr"),
-        Pair(LANGUAGE_GERMAN, "de"),
-        Pair(LANGUAGE_RUSSIAN, "ru"),
-    )
 
     override val idPrefixUri = "gallery"
